fix multiple paths/links that point to one dependency
fixes this situation:

* root binary references a symbolic link that resolves to library A
* root binary references library B
* library B references the full, "real" path to library A

this should NOT be relied on for new work because I didn't really take
the time to fully understand how it's suppose to work. but, it appears
the symbolic link feature wouldn't have worked, so I leveraged that
system in order to handle more than one path to the same library.